Create an Admin controller with 5 views   script/generate controller Admin index show new edit reset
Modify the User model to make the :role attribute accessible
  attr_accessible :role

Create the other 3 RESTful actions in the Admin controller
  :create, :update and :destroy

Write the code for all of the above actions.  The 'reset' action
is effectively the same as the edit action, but only allows editing
the password field.

Create the 5 views under app/views/admin/ :
  index.html.erb, show.html.erb, new.html.erb, reset.html.erb and edit.html.erb

add a RESTful route to routes.rb.  Include the :reset action
  map.resources :admin, :member => { :reset => [:get, :put] }