Add a @game instance variable to capture the specific game's name from the URI that was used to arrive at the controller and then pass it to the view.
  @game = params[:game] ||= ""

If the parameter is empty, we need to initialize it to an empty string.

This allows the game's name to be used in the view, not the least of
which is important for the new overridden _path helper methods.

Also, the helper methods are only visible to the views, so an path
redirected to in the controller is not working correctly at this time.
This problem will be fixed in a future update.